+++ src/InsightLogger.Api/Program.cs
@@ -6,6 +6,7 @@ using InsightLogger.Application.DependencyInjection;
 using InsightLogger.Infrastructure.DependencyInjection;
 
 var builder = WebApplication.CreateBuilder(args);
+const string AllowAllCorsPolicy = "AllowAll";
 
 builder.Logging.ClearProviders();
 builder.Logging.Configure(options =>
@@ -23,6 +24,16 @@ builder.Logging.AddJsonConsole(options =>
 
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+builder.Services.AddCors(options =>
+{
+    options.AddPolicy(AllowAllCorsPolicy, policy =>
+    {
+        // TODO: Restrict origins for non-development environments.
+        policy.AllowAnyOrigin()
+            .AllowAnyMethod()
+            .AllowAnyHeader();
+    });
+});
 builder.Services.AddInsightLoggerApi(builder.Configuration);
 builder.Services.AddInsightLoggerApplication();
 builder.Services.AddInsightLoggerInfrastructureParsing();
@@ -35,6 +46,7 @@ app.UseMiddleware<RequestLoggingMiddleware>();
 app.UseMiddleware<RequestTelemetryMiddleware>();
 app.UseMiddleware<ApiExceptionHandlingMiddleware>();
 app.UseMiddleware<RequireJsonContentTypeMiddleware>();
+app.UseCors(AllowAllCorsPolicy);
 
 app.UseSwagger(options =>
 {